Moved actions/sagas into a redux folder and cleaned up the wallet index
